2025_tinkoff_ctf/web/boomer/sources/government-biometrics/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
jinja2==3.1.2
psycopg2-binary==2.9.6
python-multipart==0.0.6 